[Ide] Move CommandManager initialization to IdeCommandManager
