+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.csproj
@@ -4243,6 +4243,7 @@
     <Compile Include="MonoDevelop.Ide.Gui.Documents\ContentCallbackRegistry.cs" />
     <Compile Include="MonoDevelop.Ide.RoslynServices\MonoDevelopExtensionManager.cs" />
     <Compile Include="MonoDevelop.Ide.RoslynServices\MonoDevelopErrorLoggerService.cs" />
+    <Compile Include="MonoDevelop.Components.Commands\IdeCommandManager.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(Configuration)' == 'DebugMac' OR '$(Configuration)' == 'ReleaseMac'">
     <Compile Include="MonoDevelop.Components\Mac\KeyCodes.cs" />
